requirements.txt
matplotlib>=3.5.0
pandas>=1.3.0
seaborn>=0.11.0
numpy>=1.21.0
scikit-learn>=0.24.0
imbalanced-learn>=0.8.0
shap>=0.40.0
joblib>=1.0.0
requests>=2.26.0
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.0
python-dotenv>=0.19.0 